+++ src/standalone/index.js
@@ -42,6 +42,7 @@ function autocomplete(selector, options, datasets, typeaheadObject) {
       minLength: options.minLength,
       autoselect: options.autoselect,
       autoselectOnBlur: options.autoselectOnBlur,
+      tabAutocomplete: options.tabAutocomplete,
       openOnFocus: options.openOnFocus,
       templates: options.templates,
       debug: options.debug,