+++ test/unit/typeahead_spec.js
@@ -94,6 +94,23 @@ describe('Typeahead', function() {
       var that = this;
       waitsForAndRuns(function() { return that.dropdown.close.calls.count(); }, done, 100);
     });
+
+    it('should pass the selection method as part of the context ', function(done) {
+      var spy;
+
+      this.$input.on('autocomplete:selected', spy = jasmine.createSpy());
+      this.dropdown.trigger('suggestionClicked');
+
+      expect(spy).toHaveBeenCalledWith(
+        jasmine.any(Object),
+        undefined,
+        undefined,
+        { selectionMethod: 'click' }
+      );
+
+      var that = this;
+      waitsForAndRuns(function() { return that.dropdown.close.calls.count(); }, done, 100);
+    });
   });
 
   describe('when dropdown triggers suggestionClicked with undefined displayKey', function() {
@@ -339,6 +356,23 @@ describe('Typeahead', function() {
       expect(this.input.setQuery).toHaveBeenCalledWith(fixtures.data.simple[1].value);
       expect(this.input.setInputValue).toHaveBeenCalledWith(fixtures.data.simple[1].value, true);
     });
+
+    it('should pass the selectionMethod as part of the context', function() {
+      this.view.autoselectOnBlur = true;
+      this.dropdown.getDatumForTopSuggestion.and.returnValue(testDatum);
+
+      var spy;
+
+      this.$input.on('autocomplete:selected', spy = jasmine.createSpy());
+      this.input.trigger('blurred');
+
+      expect(spy).toHaveBeenCalledWith(
+        jasmine.any(Object),
+        undefined,
+        undefined,
+        { selectionMethod: 'blur' }
+      );
+    });
   });
 
   describe('when debug flag is set', function() {
@@ -392,7 +426,7 @@ describe('Typeahead', function() {
         handleObj: jasmine.objectContaining({
           type: 'autocomplete:selected'
         })
-      }), undefined, undefined);
+      }), undefined, undefined, jasmine.any(Object));
       expect(view.input.setQuery).toHaveBeenCalledWith('');
       expect(view.input.setInputValue).toHaveBeenCalledWith('', true);
     });
@@ -425,8 +459,27 @@ describe('Typeahead', function() {
 
       $e = jasmine.createSpyObj('event', ['preventDefault']);
       this.input.trigger('enterKeyed', $e);
-
       expect($e.preventDefault).toHaveBeenCalled();
+    });
+
+    it('should pass the selection method as part of the context ', function(done) {
+      var $e;
+      var spy;
+      var anything = jasmine.any(Object);
+
+      $e = jasmine.createSpyObj('event', ['preventDefault']);
+      this.$input.on('autocomplete:selected', spy = jasmine.createSpy());
+      this.input.trigger('enterKeyed', $e);
+
+      expect(spy).toHaveBeenCalledWith(
+        anything,
+        undefined,
+        undefined,
+        {selectionMethod: 'enterKey'}
+      );
+
+      var that = this;
+      waitsForAndRuns(function() { return that.dropdown.close.calls.count(); }, done, 100);
     });
   });
 
@@ -461,10 +514,29 @@ describe('Typeahead', function() {
 
         expect($e.preventDefault).toHaveBeenCalled();
       });
+
+      it('should pass the selectionMethod as part of the context', function(done) {
+        var $e;
+        var spy;
+
+        $e = jasmine.createSpyObj('event', ['preventDefault']);
+        this.$input.on('autocomplete:selected', spy = jasmine.createSpy());
+        this.input.trigger('tabKeyed', $e);
+
+        expect(spy).toHaveBeenCalledWith(
+          jasmine.any(Object),
+          undefined,
+          undefined,
+          {selectionMethod: 'tabKey'}
+        );
+
+        var that = this;
+        waitsForAndRuns(function() { return that.dropdown.close.calls.count(); }, done, 100);
+      });
     });
 
     describe('when cursor is not in use', function() {
-      it('should autocomplete', function() {
+      it('should autocomplete if tabAutocomplete is true', function() {
         var spy;
 
         this.input.getQuery.and.returnValue('bi');
@@ -477,6 +549,34 @@ describe('Typeahead', function() {
 
         expect(this.input.setInputValue).toHaveBeenCalledWith(testDatum.value);
         expect(spy).toHaveBeenCalled();
+      });
+
+      it('should not autocomplete if tabAutocomplete is false', function() {
+        this.view.tabAutocomplete = false;
+        var spy;
+
+        this.input.getQuery.and.returnValue('bi');
+        this.input.getHint.and.returnValue(testDatum.value);
+        this.input.isCursorAtEnd.and.returnValue(true);
+        this.dropdown.getDatumForTopSuggestion.and.returnValue(testDatum);
+        this.$input.on('autocomplete:autocompleted', spy = jasmine.createSpy());
+
+        this.input.trigger('tabKeyed');
+
+        expect(this.input.setInputValue).not.toHaveBeenCalledWith(testDatum.value);
+        expect(spy).not.toHaveBeenCalled();
+      });
+
+      it('should close the dropdown if tabAutocomplete is false', function() {
+        this.view.tabAutocomplete = false;
+
+        this.input.getQuery.and.returnValue('bi');
+        this.input.getHint.and.returnValue(testDatum.value);
+        this.input.isCursorAtEnd.and.returnValue(true);
+
+        this.input.trigger('tabKeyed');
+
+        expect(this.dropdown.close).toHaveBeenCalled();
       });
     });
   });